Add support for autogenerated stacks
AutoScaling Groups in stacks that have been generated as an ```AWS::CloudFormation::Stack``` have a different naming scheme from the ones we had so far. The name can be _backoffice-development-AppStack-IVSLSUKVAJUW-WorkerAutoScalingGroup-1Q0HGTXVB5V4L_ which is more complex than we had before.

The updated regex should accommodate for these names as well.